requirements.txt
numpy
matplotlib
h5py
pandas
librosa
numba==0.43.0
llvmlite==0.32.1
keras==2.1.6
tensorflow-gpu==1.14.0